pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "brightsign-stuff"
requires-python = ">=3.7"
dependencies = [
    "aiohttp",
    "aiohttp-jinja2",
    "python-dotenv",
    "loguru",
    "pyquery",
    "aiofiles",
    "json-object-factory",
    "tzdata; platform_system=='Windows'",
    "click>=8.1.8",
]
version = "0.0.1"

[project.scripts]
brightsign = "brightsignweb.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"brightsignweb" = ["*.html"]
"brightsignweb.meetings" = ["*.js", "*.css", "*.svg"]
"brightsignweb.veterans" = ["*.js", "*.css", "*.svg"]
"brightsignweb.weather2" = ["*.js", "*.css", "*.svg"]

[dependency-groups]
dev = [
    "ipython>=7.34.0",
]
